change keyboard  Recently I've wrote a blog post about this exact problem you've described and how to solve it with a short and elegant way by using keyboard notifications but without using the inputAccessoryView. And although this question is pretty old this topic is still relevant so here is the link to the post: Synchronizing rotation animation between the keyboard and the attached view
If you don't want to dive into the long explanation described in the
blog post here is a short description with a code example:

The basic principle is to use the same method that everyone uses -
observing keyboard notifications to animate the attached view up and
down. But in addition to that, you have to cancel these animations when
the keyboard notifications are fired as a consequence of interface
orientation change.